Remove CheckArguments methods from Function and its sub-classes.
