+++ Functions/MinimumFiniteFunction.cs
@@ -43,20 +43,6 @@ namespace MetaphysicsIndustries.Solus.Functions
             "NaN are ignored. If no finite numbers are given, NaN is " +
             "returned.";
 
-        public override void CheckArguments(IMathObject[] args)
-        {
-            if (args.Length < 1)
-                throw new ArgumentException("No arguments passed");
-            for (var i = 0; i < args.Length; i++)
-            {
-                var argtype = args[i].GetMathType();
-                if (argtype != Reals.Value)
-                    throw new ArgumentException(
-                        $"Argument {i} wrong type: expected " +
-                        $"Scalar but got {argtype.DisplayName}");
-            }
-        }
-
         public override ISet GetResultType(SolusEnvironment env,
             IEnumerable<ISet> argTypes) => Reals.Value;
     }
